Completed add-random (the difficult version)
This version samples Unport for sequences that match the
average sequence length of an MSA and inserts as many of
those and needed to get close to the desired random ratio.
The actual random ratio is bound to be less then the
desired ratio by within 0.05 (5%).

The simple version is not done yet,
it will make duplicates of sequences in an MSA and mutate residues
in one of the two duplicates.