Fix: Real ID50s. Much smaller now.
Earlier we were taking sequences that have more than 19 mates with %id .5
but that pick up a lot of other sequences.

Now I added another step that removes those other sequences so that in the
end results all members in the group have are ID50 to one another.

This changes the results a lot.